requirements.txt
pillow
pytesseract
python-docx
openpyxl
ezodf
pdf2image